cloud_function/requirements.txt
functions-framework==3.4.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
sqlalchemy==1.4.23
psycopg2-binary==2.9.9
